feat: make config more similar to RBAC PolicyRules
This commit makes the scanner's configuration adhere to the RBAC's
"PolicyRule" type which is used within (Cluster)Roles. Without this, the
user would have needed to specify different information for the
scanner's config and the RBAC config, resulting in duplication. Instead,
the user can now simply write a slightly enhanced RBAC-style config
(having an additional "Namespaces" and "Versions" field), which is then
easily translated to the ClusterRole and scanner config.