feat: improve APIVersion selection
This commit improves the APIVersion selection for cases where the user
has asked us to scan for the preferred version, but not all resource
types might exist at that version. So far, we've then skipped all
resources that are not present in the preferredVersion.

With these changes, we'll fall back to searching all other versions as
well.