chore: add NOTICE ad bundle licenses and ensure they're up-to-date
this commit adds all licenses from our dependencies, which has been done
with the `go-licenses` tool. Additionally, it adds a `make
generate-licenses` step to easily re-run this command and regenerate
that folder.

The `NOTICE` file simply links to that folder.

Also, a CI check ensures that this folder is always up-to-date.