feat: support printing license info
This commit adds support for printing licensing information with the
`-licenses` flag. The licenses are embedded in the final binary through
a `//go:embed` directive, which allows us to read all licenses at
runtime.

The impact in binary size seems to be smaller than expected. For
reference, the `licenses` folder takes 536K on disk. However, comparing
the old vs new binary sizes shows only a difference of ~200K. I assume
the embedded FS is being compressed, but haven't found any documentation
on that.

I had to move the licenses into a separate folder because the
license-generator removed the `licenses` folder, which made the Go
module couldn't be built anymore (because the Go file was there).
Additionally, Go only allows `go:embed` directives on files / paths that
exist, so I needed to embed a path that the license-tool doesn't remove.