fix: do not crash on inexistent group
This commit fixes the config-parser to not return an error if a group
does not exist. So far we didn't return a Kubernetes "NotFound" error,
but a custom one, which meant that the matching didn't work.

We "fixed" this for the test-mock once already, but never fixed the
actual implementation...