test: helm chart config map parsing
This commit changes the existing "RealAPI" config test from using an
inline config to using the config that we template in the Helm chart.

To achieve this, we introduce a "TemplateChart" function that templates
a Helm chart at a given path with some values and returns the rendered
manifests. We then extract the ConfigMap from there and write it to a
file, which we then run our tests on.

The test assures that some default types from the Kubernetes APIs are
present. It does not test any CRDs, as we don't register any and as such
the GVK-lookup would fail.