feat: behavior & documentation on namespaces setting
This commit fixes the documentation on the `namespaces` key in the scan
types. The comment previously said that a 0-length list (`[]`) has the
same behavior as `null`. However, that wasn't actually true, and we even
had this bug in the Helm chart.

To fix this, I could have also changed the code to treat null and empty
as the same. However, I decided to differentiate it so that an empty
list can be used to skip all namespaced resources: by setting `[]`, it
will skip all namespaces and thus only reconcile cluster-scoped
resources. While this isn't particularly useful today, this could change
if we'd ever introduce the `resources: ["*"]` setting / feature.

(One could argue that we could have changed the code at that point as
well, but that would have introduced a breaking change, which we can
avoid now).

This commit is marked as "feat" because `semantic-release` does not
handle `v0` SemVer versions. Technically, this commit is a breaking
change because it changes the behavior of the config. However, we're
`v0` and I only want to bump to the next major version (e.g. `v0.4.0` ->
`v0.5.0`), as is normal in SemVer to indicate breaking changes in
alpha-software. Just a fix-bump wouldn't have captured the breaking
change.