feat: add failure age metric
This commit adds another metric that tracks the *age* of the latest
failure, instead of the timestamp.

We're adding this because Datadog doesn't seem to have a `time()`
function like prometheus, so subtracting `now - latest-failure` isn't
easily possible in Datadog. Having a gauge in its place that simply
indicates the age is a lot easier.