chore: use new resource identifier to store failures
We've noticed that we have a dangling resource in our failure-map in the
dev-enviroment. The underlying reason for this dangling resource is
that the call to our API failed when the resource was being deleted, and
because it then never successfully reconciled again, it wasn't cleaned
up from the failure map.

The issue has two root causes:
1) When resources are in deletion and the API call fails, we
   `recordFailure` nonetheless. Because the resource will not be
    reconciled at all anymore, we have no chance to `recordSuccess`,
    so the resource will be abandoned and stay in the failure map
    forever.
2) We identify resources within that failure map by UID. Within a
   deletion call, that UID will be empty. As such, even if we wanted to
    remove a resource from the failure map at it's deletion call, we
    couldn't locate it in the map.

This commit fixes 2), identifying resources by GVKNN instead of the UID.
This also lays the groundwork for fixing 1).