fix: remove resources from failure map on deletion calls
We've noticed that we have a dangling resource in our failure-map in the
dev-enviroment. The underlying reason for this dangling resource is
that the call to our API failed when the resource was being deleted, and
because it then never successfully reconciled again, it wasn't cleaned
up from the failure map.

The issue has two root causes:
1) When resources are in deletion and the API call fails, we
   `recordFailure` nonetheless. Because the resource will not be
    reconciled at all anymore, we have no chance to `recordSuccess`,
    so the resource will be abandoned and stay in the failure map
    forever.
2) We identify resources within that failure map by UID. Within a
   deletion call, that UID will be empty. As such, even if we wanted to
    remove a resource from the failure map at it's deletion call, we
    couldn't locate it in the map.

Number 2) has been fixed in a previous commit, while this commit fixes
1). If we do not get a UID - which signals that the resource is being
deleted - we do not record the failure of the resource and instead call
`recordSuccess`. While technically the call wasn't a success,
`recordSuccess` does exactly what we want it to do: remove the resource
from the failure map and decrease any required counters.