fix: oldest failure age metric when there are no failures
Prior to this commit, we set this metric to +Inf, because semantically,
the last error never happened. In practice, this is awkward. Following
the recommendations in monitoring.md, a value of +Inf that actually
represents a perfectly healthy state, would result in an alert that
always fires.

Using zero instead is perhaps a little confusing, but better expresses
what we care about from an alerting perspective: there is no persistent
reconciliation failure.