test: fix flaky smoke test
Our Smoke Tests were _slightly_ flaky (~1%?), due to the logic in our
smoke test.

What happened in one of these flaky runs is that the initial `NodeList`
was populated *before* all pods where deployed, and then we queried the
backend and the resources didn't match because the backend was actually
already updated!

The fix for this is to get the nodes before every `List` call to the
backend, which makes sure we include such updates.