fix: sanity check for credential only [IA-360] (#79)
In #77, I added a sanity check on startup. However, because this
requests `/orgs/{org_id}`, this sanity check fails in the case a
customer is using a custom role that **only** has permissions to publish
kubernetes resources.

This PR changes that sanity check to use `/self` instead, which should
always be available.